Commit para deploy. Alguns detalhes tem de ser testados, como retorno de telas, login com sucesso, inserção de dados....

+++ src/main/java/br/com/gestaonotavel/ifsul/dao/PacienteDAO.java
@@ -2,7 +2,6 @@ package br.com.gestaonotavel.ifsul.dao;
 
 import br.com.gestaonotavel.ifsul.model.Paciente;
 import br.com.gestaonotavel.ifsul.util.JpaUtil;
-
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
 import javax.persistence.NoResultException;
@@ -15,8 +14,6 @@ public class PacienteDAO {
         EntityManager em = JpaUtil.getEntityManager();
         EntityTransaction tx = em.getTransaction();
         tx.begin();
-
-
         try {
             Paciente pacienteSalvo = em.merge(paciente);
             tx.commit();
@@ -29,21 +26,10 @@ public class PacienteDAO {
         }
     }
 
-    public Paciente buscarPorId(Long id) {
-        EntityManager em = JpaUtil.getEntityManager();
-        try {
-            return em.find(Paciente.class, id);
-        } finally {
-            em.close();
-        }
-    }
-
     public List<Paciente> listarTodos() {
         EntityManager em = JpaUtil.getEntityManager();
         try {
             return em.createQuery("SELECT DISTINCT p FROM Paciente p LEFT JOIN FETCH p.responsaveisLista", Paciente.class).getResultList();
-        } catch (PersistenceException ex) {
-            throw ex;
         } finally {
             em.close();
         }
@@ -55,14 +41,10 @@ public class PacienteDAO {
         tx.begin();
         try {
             Paciente paciente = em.find(Paciente.class, id);
-            if (paciente != null) {
-                em.remove(paciente);
-            }
+            if (paciente != null) em.remove(paciente);
             tx.commit();
         } catch (PersistenceException ex) {
-            if (tx.isActive()) {
-                tx.rollback();
-            }
+            if (tx.isActive()) tx.rollback();
             throw ex;
         } finally {
             em.close();
@@ -73,11 +55,10 @@ public class PacienteDAO {
         EntityManager em = JpaUtil.getEntityManager();
         try {
             return em.createQuery("SELECT p FROM Paciente p WHERE p.cpf = :cpf", Paciente.class)
-                    .setParameter("cpf", cpfBuscado)
-                    .getSingleResult();
-        }catch (NoResultException e){
+                    .setParameter("cpf", cpfBuscado).getSingleResult();
+        } catch (NoResultException e) {
             return null;
-        }finally {
+        } finally {
             em.close();
         }
     }
